C3_Longest_Substring_Without_Repeating_Characters/c3_test.go
package c3

import "testing"

func Test_lengthOfLongestSubstring(t *testing.T) {
	type args struct {
		s string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "Example 0",
			args: args{
				s: "abcabcbb",
			},
			want: 3,
		},
		{
			name: "Example 1",
			args: args{
				s: "bbbbbbbbb",
			},
			want: 1,
		},
		{
			name: "Example 2",
			args: args{
				s: "pwwkew",
			},
			want: 3,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := lengthOfLongestSubstring(tt.args.s); got != tt.want {
				t.Errorf("lengthOfLongestSubstring() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_lengthOfLongestSubstringMap(t *testing.T) {
	type args struct {
		s string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "Example 0",
			args: args{
				s: "abcabcbb",
			},
			want: 3,
		},
		{
			name: "Example 1",
			args: args{
				s: "bbbbbbbbb",
			},
			want: 1,
		},
		{
			name: "Example 2",
			args: args{
				s: "pwwkew",
			},
			want: 3,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := lengthOfLongestSubstringMap(tt.args.s); got != tt.want {
				t.Errorf("lengthOfLongestSubstring() = %v, want %v", got, tt.want)
			}
		})
	}
}
